requirements.txt
certifi==2019.9.11
chardet==3.0.4
Deprecated==1.2.6
idna==2.8
pathlib==1.0.1
PyGithub==1.43.8
PyJWT==1.7.1
PyYAML==5.1.2
requests==2.22.0
urllib3==1.25.6
wrapt==1.11.2
